SAK-508 repository refactoring
