+++ src/Snt/Capi/Repository/Article/ArticleRepository.php
@@ -3,43 +3,24 @@
 namespace Snt\Capi\Repository\Article;
 
 use Snt\Capi\Http\Exception\HttpException;
-use Snt\Capi\Http\HttpClientInterface;
-use Snt\Capi\Repository\Exception\CouldNotFetchResourceRepositoryException;
+use Snt\Capi\Repository\AbstractRepository;
 use Snt\Capi\Repository\FindParametersInterface;
 
-class ArticleRepository implements ArticleRepositoryInterface
+class ArticleRepository extends AbstractRepository implements ArticleRepositoryInterface
 {
-    /**
-     * @var HttpClientInterface
-     */
-    protected $httpClient;
-
-    /**
-     * @param HttpClientInterface $httpClient
-     */
-    public function __construct(
-        HttpClientInterface $httpClient
-    ) {
-        $this->httpClient = $httpClient;
-    }
-
     /**
      * {@inheritdoc}
      */
     public function find(FindParametersInterface $findParameters)
     {
         try {
-            $articleRawData = $this->makeRequest($findParameters);
+            $articleRawData = json_decode($this->makeHttpGetRequest($findParameters), true);
         } catch (HttpException $exception) {
             if ($exception->isNotFoundError()) {
                 return null;
             }
 
-            throw new CouldNotFetchResourceRepositoryException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                $exception
-            );
+            $this->throwCouldNotFetchException($exception);
         }
 
         return $articleRawData;
@@ -51,17 +32,9 @@ class ArticleRepository implements ArticleRepositoryInterface
     public function findByIds(FindParametersInterface $findParameters)
     {
         try {
-            $articlesRawData = $this->makeRequest($findParameters);
+            $articlesRawData = json_decode($this->makeHttpGetRequest($findParameters), true);
         } catch (HttpException $exception) {
-            if ($exception->isNotFoundError()) {
-                return null;
-            }
-
-            throw new CouldNotFetchResourceRepositoryException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                $exception
-            );
+            $this->throwCouldNotFetchException($exception);
         }
 
         return isset($articlesRawData['articles']) ? $articlesRawData['articles'] : [];
@@ -73,25 +46,11 @@ class ArticleRepository implements ArticleRepositoryInterface
     public function findByChangelog(FindParametersInterface $findParameters)
     {
         try {
-            $articlesChangelogRawData = $this->makeRequest($findParameters);
+            $articlesChangelogRawData = json_decode($this->makeHttpGetRequest($findParameters), true);
         } catch (HttpException $exception) {
-            throw new CouldNotFetchResourceRepositoryException(
-                $exception->getMessage(),
-                $exception->getCode(),
-                $exception
-            );
+            $this->throwCouldNotFetchException($exception);
         }
 
         return $articlesChangelogRawData;
-    }
-
-    private function makeRequest(FindParametersInterface $findParameters)
-    {
-        return json_decode(
-            $this->httpClient->get(
-                $findParameters->buildHttpRequestParameters()
-            ),
-            true
-        );
     }
 }
